terraform/src/cloud_functions/get_audio_query/requirements.txt
google-auth==2.16.2
requests==2.27.1